refactor: 레포지터리 경로 변경 및 redis 관련 기능 이식중
