+++ src/main/java/com/st/eighteen_be/chat/api/ChattingApiController.java
@@ -6,8 +6,8 @@ import com.st.eighteen_be.chat.model.dto.request.EnterChatRoomRequestDTO;
 import com.st.eighteen_be.chat.model.dto.request.FindChatRoomRequestDTO;
 import com.st.eighteen_be.chat.model.dto.response.ChatMessageResponseDTO;
 import com.st.eighteen_be.chat.model.dto.response.ChatroomWithLastestMessageDTO;
+import com.st.eighteen_be.chat.service.ChatroomService;
 import com.st.eighteen_be.chat.service.facade.ChatroomFacade;
-import com.st.eighteen_be.chat.service.impl.ChatroomService;
 import com.st.eighteen_be.chat.service.kafka.ChattingProducer;
 import com.st.eighteen_be.common.response.ApiResponse;
 import jakarta.validation.Valid;
@@ -40,29 +40,29 @@ public class ChattingApiController {
     private final ChattingProducer chattingProducer;
     private final ChatroomFacade chatroomFacade;
     private final ChatroomService chatroomService;
-    
+
     @MessageMapping("/chat/all")
     @ResponseStatus(HttpStatus.OK)
     public ApiResponse<List<ChatroomWithLastestMessageDTO>> findAllMyChatrooms(@Valid @RequestBody FindChatRoomRequestDTO requestDTO) {
         log.info("findAllMyChatrooms.requestDTO.senderNo() = {}", requestDTO.senderNo());
-        
+
         return ApiResponse.success(HttpStatus.OK, chatroomService.findAllMyChatrooms(requestDTO));
     }
-    
+
     @MessageMapping("/chat/enter") // /pub/chat/enter
     public ApiResponse<List<ChatMessageResponseDTO>> enterChatroom(@Valid @RequestBody EnterChatRoomRequestDTO requestDTO) {
         log.info("enterChatroom.requestDTO.senderNo() = {}", requestDTO.senderNo());
         log.info("enterChatroom.requestDTO.receiverNo() = {}", requestDTO.receiverNo());
         log.info("enterChatroom.requestDTO.requestTime() = {}", requestDTO.requestTime());
-        
+
         return ApiResponse.success(HttpStatus.OK, chatroomFacade.getOrCreateChatroom(requestDTO));
     }
-    
+
     @MessageMapping("/chat/{senderNo}/{receiverNo}/message") // /pub/chat/{chatroomId}/message
     public void sendMessage(@DestinationVariable(value = "senderNo") Long senderNo, @DestinationVariable(value = "receiverNo") Long receiverNo, @RequestBody ChatMessageRequestDTO chatMessage) {
         log.info("sendMessage.senderNo() = {}, receiverNo() = {}", senderNo, receiverNo);
         log.info("sendMessage.chatMessage.message() = {}", chatMessage.getMessage());
-        
+
         chattingProducer.send(KafkaConst.CHAT_TOPIC, chatMessage);
     }
 }
