+++ src/main/java/com/st/eighteen_be/chat/service/ChatMessageService.java
@@ -1,4 +1,4 @@
-package com.st.eighteen_be.chat.service.impl;
+package com.st.eighteen_be.chat.service;
 
 import com.st.eighteen_be.chat.model.collection.ChatMessageCollection;
 import com.st.eighteen_be.chat.model.dto.request.ChatMessageRequestDTO;
@@ -23,46 +23,46 @@ import java.util.List;
 @RequiredArgsConstructor
 @Slf4j
 public class ChatMessageService {
-    
+
     private final ChatMessageCollectionRepository chatMessageCollectionRepository;
     private final ChatroomInfoCollectionRepository chatroomInfoCollectionRepository;
-    
+
     @Transactional(readOnly = false)
     public void processMessage(ChatMessageRequestDTO messageDto) {
         log.info("========== processMessage ========== senderNo : {}, receiverNo : {}", messageDto.getSenderNo(), messageDto.getReceiverNo());
-        
+
         ChatMessageCollection chatMessage = messageDto.toCollection();
-        
+
         chatroomInfoCollectionRepository.findBySenderNoAndReceiverNo(chatMessage.getSenderNo(), chatMessage.getReceiverNo())
                 .ifPresentOrElse(
                         chatroomInfo -> {
                             log.info("========== chatroom found ==========");
-                            
+
                             addMessage(chatMessage);
                         },
                         () -> {
                             log.info("========== chatroom not found ==========");
-                            
+
                             throw new NotFoundException(ErrorCode.NOT_FOUND_CHATROOM);
                         }
                 );
     }
-    
+
     public List<ChatMessageResponseDTO> findMessagesBeforeTimeInRoom(Long senderNo, Long receiverNo, LocalDateTime lastMessageTime) {
         log.info("========== findMessagesBeforeTimeInRoom ========== senderNo : {}, receiverNo : {}, lastMessageTime : {}", senderNo, receiverNo, lastMessageTime);
-        
+
         Pageable pageable = PageRequest.of(0, 20, Sort.by("createdAt").descending());
-        
+
         List<ChatMessageCollection> foundChatMessages = chatMessageCollectionRepository.findBySenderNoAndReceiverNoAndCreatedAtBefore(senderNo, receiverNo, lastMessageTime, pageable);
-        
+
         return foundChatMessages.stream()
                 .map(ChatMessageCollection::toResponseDTO)
                 .toList();
     }
-    
+
     private void addMessage(ChatMessageCollection chatMessage) {
         log.info("========== addMessage ==========");
-        
+
         chatMessageCollectionRepository.save(chatMessage);
     }
 }
