+++ src/main/java/com/st/eighteen_be/chat/service/kafka/ChattingConsumer.java
@@ -2,7 +2,7 @@ package com.st.eighteen_be.chat.service.kafka;
 
 import com.st.eighteen_be.chat.constant.KafkaConst;
 import com.st.eighteen_be.chat.model.dto.request.ChatMessageRequestDTO;
-import com.st.eighteen_be.chat.service.impl.ChatMessageService;
+import com.st.eighteen_be.chat.service.ChatMessageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -28,17 +28,17 @@ import java.text.MessageFormat;
 @RequiredArgsConstructor
 @Slf4j
 public class ChattingConsumer {
-    
+
     private final SimpMessagingTemplate messagingTemplate;
     private final ChatMessageService chatMessageService;
-    
+
     @Transactional(readOnly = false)
     @KafkaListener(topics = KafkaConst.CHAT_TOPIC, groupId = KafkaConst.CHAT_CONSUMER_GROUP_ID)
     public void listen(ChatMessageRequestDTO messageDto) {
         log.info("ChattingConsumer.listen :  senderNo={}, receiverNo={}, message={}, chatroomInfo={}: ", messageDto.getSenderNo(), messageDto.getReceiverNo(), messageDto.getMessage(), messageDto.getChatroomInfoId());
-        
+
         messagingTemplate.convertAndSend(MessageFormat.format("/sub/chat/{0}/{1}/message", messageDto.getSenderNo(), messageDto.getReceiverNo()), messageDto);
-        
+
         chatMessageService.processMessage(messageDto);
     }
 }
