+++ src/main/java/com/st/eighteen_be/chat/service/kafka/ChattingProducer.java
@@ -2,7 +2,7 @@ package com.st.eighteen_be.chat.service.kafka;
 
 import com.st.eighteen_be.chat.model.collection.ChatroomInfoCollection;
 import com.st.eighteen_be.chat.model.dto.request.ChatMessageRequestDTO;
-import com.st.eighteen_be.chat.service.impl.ChatroomService;
+import com.st.eighteen_be.chat.service.ChatroomService;
 import com.st.eighteen_be.common.exception.ErrorCode;
 import com.st.eighteen_be.common.exception.sub_exceptions.data_exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -25,23 +25,23 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 @Slf4j
 public class ChattingProducer {
-    
+
     private final KafkaTemplate<String, ChatMessageRequestDTO> kafkaTemplate;
     private final ChatroomService chatroomService;
-    
+
     public void send(String topic, ChatMessageRequestDTO messageDto) {
         log.info("ChattingProducer.sendTopic : {}", topic);
         log.info("ChattingProducer.messageContent : senderNo={}, receiverNo={}, message={}: ", messageDto.getSenderNo(), messageDto.getReceiverNo(), messageDto.getMessage());
-        
+
         setChatroomInfo(messageDto);
-        
+
         kafkaTemplate.send(topic, messageDto);
     }
-    
+
     private void setChatroomInfo(ChatMessageRequestDTO messageDto) {
         ChatroomInfoCollection foundedChatroom = chatroomService.getChatroom(messageDto.getSenderNo(), messageDto.getReceiverNo())
                 .orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_CHATROOM));
-        
+
         messageDto.setChatroomInfoId(foundedChatroom.get_id());
     }
 }
