+++ src/main/java/com/st/eighteen_be/chat/service/redis/RedisMessageService.java
@@ -1,0 +1,23 @@
+package com.st.eighteen_be.chat.service.redis;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+/**
+ * packageName    : com.st.eighteen_be.chat.service.redis
+ * fileName       : RedisMessageService
+ * author         : ipeac
+ * date           : 24. 5. 10.
+ * description    :
+ * ===========================================================
+ * DATE              AUTHOR             NOTE
+ * -----------------------------------------------------------
+ * 24. 5. 10.        ipeac       최초 생성
+ */
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class RedisMessageService {
+    private final springredis
+}
