+++ src/test/java/com/st/eighteen_be/chat/service/kafka/ChattingKafkaTest.java
@@ -5,7 +5,7 @@ import com.st.eighteen_be.chat.model.collection.ChatroomInfoCollection;
 import com.st.eighteen_be.chat.model.dto.request.ChatMessageRequestDTO;
 import com.st.eighteen_be.chat.model.vo.ChatroomType;
 import com.st.eighteen_be.chat.repository.mongo.ChatroomInfoCollectionRepository;
-import com.st.eighteen_be.chat.service.impl.ChatroomService;
+import com.st.eighteen_be.chat.service.ChatroomService;
 import com.st.eighteen_be.common.annotation.ServiceWithMongoDBTest;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
@@ -43,24 +43,24 @@ import static org.assertj.core.api.Assertions.assertThat;
 @ExtendWith(MockitoExtension.class)
 @Testcontainers
 public class ChattingKafkaTest {
-    
+
     @Container
     private KafkaContainer kafkaContainer = new KafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:7.6.0")).withKraft()
             .withExposedPorts(9093);
-    
+
     @Autowired
     private ChatroomInfoCollectionRepository chatroomInfoCollectionRepository;
-    
+
     private KafkaTemplate<String, ChatMessageRequestDTO> kafkaTemplate;
-    
+
     private ChatroomService chatroomService;
-    
+
     private Consumer<String, ChatMessageRequestDTO> consumer;
-    
+
     private ChattingProducer chattingProducer;
-    
+
     private ChatMessageRequestDTO messageDto;
-    
+
     @BeforeEach
     void setUp() {
         // 카프카 프로듀서 설정
@@ -68,15 +68,15 @@ public class ChattingKafkaTest {
         configs.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaContainer.getBootstrapServers());
         configs.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
         configs.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-        
+
         // ProducerFactory와 KafkaTemplate 생성
         ProducerFactory<String, ChatMessageRequestDTO> producerFactory = new DefaultKafkaProducerFactory<>(configs);
         kafkaTemplate = new KafkaTemplate<>(producerFactory);
-        
+
         chatroomService = new ChatroomService(chatroomInfoCollectionRepository);
-        
+
         chattingProducer = new ChattingProducer(kafkaTemplate, chatroomService);
-        
+
         //consumer 설정
         Properties consumerProps = new Properties();
         consumerProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaContainer.getBootstrapServers());
@@ -85,30 +85,30 @@ public class ChattingKafkaTest {
         consumerProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class.getName());
         consumerProps.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class.getName());
         consumerProps.put(JsonDeserializer.TRUSTED_PACKAGES, "*");
-        
+
         consumer = new KafkaConsumer<>(consumerProps);
         consumer.subscribe(Collections.singleton(KafkaConst.CHAT_TOPIC));
-        
+
         messageDto = ChatMessageRequestDTO.builder()
                 .senderNo(1L)
                 .receiverNo(2L)
                 .message("Hello")
                 .build();
     }
-    
+
     @Test
     @DisplayName("메시지 전송 성공 - 프로듀서 - 컨슈머 정상 동작 테스트")
     void When_SendMessage_With_Producer_Expect_Success() {
         //given
         //채팅방이 형성되어 있어야함.
         chatroomInfoCollectionRepository.save(ChatroomInfoCollection.of(1L, 2L, ChatroomType.PRIVATE));
-        
+
         // when
         chattingProducer.send(KafkaConst.CHAT_TOPIC, messageDto);
-        
+
         // then
         ConsumerRecord<String, ChatMessageRequestDTO> record = KafkaTestUtils.getSingleRecord(consumer, KafkaConst.CHAT_TOPIC);
-        
+
         assertThat(record).isNotNull();
         assertThat(record.value().getSenderNo()).isEqualTo(messageDto.getSenderNo());
         assertThat(record.value().getMessage()).isEqualTo(messageDto.getMessage());
